pyproject.toml
[project]
name = "spirit-rules-config"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "databricks-sdk>=0.91.0",
    "fastapi>=0.129.0",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
